backend_requirements.txt
fastapi
uvicorn
python-multipart
pydantic[email]
pytesseract
pdf2image
Pillow
fastapi-mail
python-dotenv